+++ stm32f103c8t6/02-demo-mimiscript-stm32f103c8t6-usart/mimiscript-core/dataString.c
@@ -1,12 +1,13 @@
 #include "dataString.h"
 #include <stdio.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 
-char *strAppendWithSize_unlimited(char *strOut, char *pData, int Size)
+char *strAppendWithSize_unlimited(char *strOut, char *pData, int32_t Size)
 {
-	int strOut_i = strGetSize(strOut);
-	for (int i = 0; i < Size; i++)
+	int32_t strOut_i = strGetSize(strOut);
+	for (int32_t i = 0; i < Size; i++)
 	{
 		strOut[strOut_i + i] = pData[i];
 	}
@@ -19,11 +20,11 @@ char *strAppendWithSize_unlimited(char *strOut, char *pData, int Size)
 
 char *strCut(char *strOut, char *strIn, char startSign, char endSign)
 {
-	int Size = strGetSize(strIn);
-	int isStart = 0;
-	int isEnd = 0;
-	int iOut = 0;
-	for (int i = 0; i < Size; i++)
+	int32_t Size = strGetSize(strIn);
+	int32_t isStart = 0;
+	int32_t isEnd = 0;
+	int32_t iOut = 0;
+	for (int32_t i = 0; i < Size; i++)
 	{
 		if (!isStart && (strIn[i] == startSign))
 		{
@@ -54,8 +55,8 @@ char *strCut(char *strOut, char *strIn, char startSign, char endSign)
 
 char *strDeleteChar(char *strOut, char *strIn, char ch)
 {
-	int iOut = 0;
-	for (int i = 0; i < strGetSize(strIn); i++)
+	int32_t iOut = 0;
+	for (int32_t i = 0; i < strGetSize(strIn); i++)
 	{
 		if (ch == strIn[i])
 		{
@@ -74,10 +75,10 @@ char *strDeleteEnter(char *str)
 	return strDeleteChar(str, str, '\n');
 }
 
-char *strAppendWithSize(char *strOut, char *pData, int Size)
+char *strAppendWithSize(char *strOut, char *pData,int32_t Size)
 {
-	unsigned short int strOut_i = strGetSize(strOut);
-	for (int i = 0; i < Size; i++)
+	int32_t strOut_i = strGetSize(strOut);
+	for (int32_t i = 0; i < Size; i++)
 	{
 		strOut[strOut_i + i] = pData[i];
 	}
@@ -88,10 +89,10 @@ char *strAppendWithSize(char *strOut, char *pData, int Size)
 	return strOut;
 }
 
-int strCountSign(char *strIn, char sign)
+int32_t strCountSign(char *strIn, char sign)
 {
-	int count = 0;
-	for (int i = 0; i < strGetSize(strIn); i++)
+	int32_t count = 0;
+	for (int32_t i = 0; i < strGetSize(strIn); i++)
 	{
 		if (sign == strIn[i])
 		{
@@ -101,19 +102,19 @@ int strCountSign(char *strIn, char sign)
 	return count;
 }
 
-int strGetTokenNum(char *strIn, char sign)
+int32_t strGetTokenNum(char *strIn, char sign)
 {
 	return strCountSign(strIn, sign) + 1;
 }
 
-unsigned short int strGetSize(char *pData)
+uint32_t strGetSize(char *pData)
 {
-	unsigned short int Size = 0;
+	uint32_t Size = 0;
 	if (NULL == pData)
 	{
 		return 0;
 	}
-	for (int i = 0; pData[i] != 0; i++)
+	for (int32_t i = 0; pData[i] != 0; i++)
 	{
 		Size++;
 	}
@@ -122,17 +123,17 @@ unsigned short int strGetSize(char *pData)
 
 char *strAppend_unlimited(char *strOut, char *pData)
 {
-	unsigned short int Size = 0;
+	uint32_t Size = 0;
 	Size = strGetSize(pData);
 	return strAppendWithSize_unlimited(strOut, pData, Size);
 }
 
 char *strGetLastToken(char *strOut, char *strIn, char sign)
 {
-	int size = strGetSize(strIn);
-	int isGetSign = 0;
-	int buffSize = 0;
-	for (int i = size - 1; i > -1; i--)
+	int32_t size = strGetSize(strIn);
+	int32_t isGetSign = 0;
+	int32_t buffSize = 0;
+	for (int32_t i = size - 1; i > -1; i--)
 	{
 		if (strIn[i] != sign)
 		{
@@ -145,7 +146,7 @@ char *strGetLastToken(char *strOut, char *strIn, char sign)
 			break;
 		}
 	}
-	int i = 0;
+	int32_t i = 0;
 	for (i = 0; i < buffSize / 2; i++)
 	{
 		char buff = strOut[i];
@@ -158,11 +159,11 @@ char *strGetLastToken(char *strOut, char *strIn, char sign)
 
 char *strPopToken(char *strOut, char *strIn, char sign)
 {
-	int getSign = 0;
-	int iPoped = 0;
-	int iOut = 0;
-	int size = strGetSize(strIn);
-	int i = 0;
+	int32_t getSign = 0;
+	int32_t iPoped = 0;
+	int32_t iOut = 0;
+	int32_t size = strGetSize(strIn);
+	int32_t i = 0;
 	for (i = 0; i < size; i++)
 	{
 		if (getSign)
@@ -188,8 +189,8 @@ char *strPopToken(char *strOut, char *strIn, char sign)
 
 char *strGetFirstToken(char *strOut, char *strIn, char sign)
 {
-	int size = strGetSize(strIn);
-	for (int i = 0; i < size; i++)
+	int32_t size = strGetSize(strIn);
+	for (int32_t i = 0; i < size; i++)
 	{
 		if (strIn[i] != sign)
 		{
@@ -203,14 +204,14 @@ char *strGetFirstToken(char *strOut, char *strIn, char sign)
 	return strOut;
 }
 
-int strGetToken(char *string, char **argv, char sign)
+int32_t strGetToken(char *string, char **argv, char sign)
 {
-	int argc = 0;
-	int i = 0;
-	//arg_i point to the arg operated now
-	int arg_i = 0;
+	int32_t argc = 0;
+	int32_t i = 0;
+	//arg_i point32_t to the arg operated now
+	int32_t arg_i = 0;
 	// if not found ' ', then put chars from CMD to argv_tem
-	int char_i = 0;
+	int32_t char_i = 0;
 	for (i = 0; (i < strGetSize(string)); i++)
 	{
 		if (string[i] != sign)
@@ -234,21 +235,21 @@ int strGetToken(char *string, char **argv, char sign)
 
 char *strAppend(char *strOut, char *pData)
 {
-	unsigned short int Size = 0;
+	uint32_t Size = 0;
 	Size = strGetSize(pData);
 	return strAppendWithSize(strOut, pData, Size);
 }
 
-int strIsStartWith(char *str, char *strStart)
+int32_t strIsStartWith(char *str, char *strStart)
 {
 	if (NULL == str || NULL == strStart)
 	{
 		/* input is null */
 		return 0;
 	}
-	unsigned short int size = strGetSize(strStart);
-	unsigned short int CMDName_get = 1;
-	for (int i = 0; i < size; i++)
+	uint32_t size = strGetSize(strStart);
+	uint32_t CMDName_get = 1;
+	for (int32_t i = 0; i < size; i++)
 	{
 		if (str[i] != strStart[i])
 		{
@@ -258,15 +259,15 @@ int strIsStartWith(char *str, char *strStart)
 	return CMDName_get;
 }
 
-int strEqu(char *str1, char *str2)
+int32_t strEqu(char *str1, char *str2)
 {
-	int size1 = strGetSize(str1);
-	int size2 = strGetSize(str2);
+	int32_t size1 = strGetSize(str1);
+	int32_t size2 = strGetSize(str2);
 	if (size1 != size2)
 	{
 		return 0; // not equal
 	}
-	for (int i = 0; i < size1; i++)
+	for (int32_t i = 0; i < size1; i++)
 	{
 		if (str1[i] != str2[i])
 		{
@@ -283,7 +284,7 @@ char *strRemovePrefix(char *inputStr, char *prefix, char *outputStr)
 		return NULL;
 	}
 
-	for (int i = strGetSize(prefix); i < strGetSize(inputStr); i++)
+	for (int32_t i = strGetSize(prefix); i < strGetSize(inputStr); i++)
 	{
 		outputStr[i - strGetSize(prefix)] = inputStr[i];
 	}
@@ -292,16 +293,16 @@ char *strRemovePrefix(char *inputStr, char *prefix, char *outputStr)
 
 char *strClear(char *str)
 {
-	for (int i = 0; i < sizeof(str); i++)
+	for (int32_t i = 0; i < sizeof(str); i++)
 	{
 		str[i] = 0;
 	}
 	return str;
 }
 
-int strIsContain(char *str, char ch)
+int32_t strIsContain(char *str, char ch)
 {
-	for (int i = 0; i < strGetSize(str); i++)
+	for (int32_t i = 0; i < strGetSize(str); i++)
 	{
 		if (str[i] == ch)
 		{
